Prevent overlapping concurrent requests.
If the server takes longer than the polling interval to process
requests, we end up builing an ever-growing backlog of outstanding
requests.

It looks like the original intention of the isProcessing variable was to
prevent this from occuring. This commit fixes the implementation so that
only a single request can be outstanding at a time.